Abstracting the app logic that exists on popup to the background
The main app logic was existing on the popup code and it was limiting
new functionalities since that all the logic only work in the context
that was only functional when the popup was open.

For example, the buttons on the notification to start or finish runnables
runs on the background context, so, without the popup code, we'd need to
recreate the whole redux structure in background to just start/finish
runnable, not to mention the fact of eventual concurrency problems.

So, move the "core" to the background and do all the state mutations orders
through messages looks like the right choice in this moment.